fix: Send final report in automatic mode :breaking:
Send final report when guessing multiple files and the --partial
flag is not specified.
In this situation the previous behaviour was to send some partial
reports but without a final at the end. This way users need to
send a manual final right after.
With this change the final is sent automatically.