doc: Clarify how to run codacy-coverage-reporter
The Codacy Coverage Reporter must run on the root of a locally checked
out branch of the Git repository.

Also, it's more explicit to recommend specifying the path to the
coverage report manually since the Coverage Reporter may not find all
coverage report files (for example, if they're in a subdirectory).