clean: Reorder troubleshooting sections
This promotes the section "No coverage data is visible on the Codacy
UI" as the starting point for users to start troubleshooting their
setup.

I also placed the section "Can't validate checksum" at the bottom of
the page, because this scenario is more unlikely to happen in the
most recent version of Codacy Coverage Reporter.