clean: Simplify coverage setup process
Checking the console output should be an integral part of the step
to run the Codacy Coverage Reporter.

Also removes the optional step to add a Codacy badge for coverage,
since at this time the metric probably won't be available yet and
the badge will appear greyed out, which can confuse the users.